Maintain branch with precompiled binaries
We keep a copy of the main branch + precompiled binaries in prod/test-binaries
so consumers can easily add the repository as a submodule pointing to this
branch. Developers who want to contribute to this repository but don't want to
pull all the binaries (their size might grow over time) can use `--single-branch`
flag to pull only the `main` branch.