Port more tests from Wasmtime's testsuite
Following up on ac32f57400cdcdd0425d3085c24fc7fc40011d1c, port several
more tests from Wasmtime's testsuite to wasi-testsuite.

It is likely that at least some of these tests are testing behavior
which differs between Wasm engine implementations. The intent here is
not to instate these as authoritative, but to help surface these
differences so that we can discuss whether it's best to change the test,
or add documentation to the spec.

Other Wasm engines are welcome to submit their tests to wasi-testsuite
as well.