Port some more wasmtime filesystem tests
To close the testing gaps on any filesystem functions, port some more of
the wasmtime tests. A few test cases in dir_fd_op_failures seemed to be
specific to wasmtime; namely fd_sync, fd_datasync, fd_fdstat_set_flags
and fd_advise. These functions return errors when called on a directory
fd on wasmtime but on WAMR (and POSIX) they return 0 so for the moment
have removed them.