Add explicit test for fd_fdstat_set_rights
Add an explicit test to verify the behaviour of fd_fdstat_set_rights. A
few cases are covered:

1. Operations on an fd with sufficient rights succeed. Dropping the
relevant rights on same fd will cause the same operations to fail due to
insufficient rights.
2. Any dropped rights are reflected in the returned rights from fd_fdstat_get.
3. Additional rights cannot be granted to an fd via fd_fdstat_rights.

The new test covers the cases (and more) previously covered by
path_open_read_without_rights so remove that test in favour of this one.