Add expected error code for fd_prestat_dir_name
In the case that the user provides a buffer which is too small to fit
the directory name, it makes more sense to return EINVAL rather than
ENAMETOOLONG. ENAMETOOLONG has the speficic connotation that a path was
longer than the maximum size permitted by the OS - that doesn't really
match this error condition. We're keeping both error codes for now since
we don't specify in the docs what error code should be returned.