Add test for pwrite with append flag
WASI specifies `fd_pwrite` to be similar to [POSIX `pwritev`][1] which
specifies that it should ignore the append flag and write from the
supplied offset.  Currently, only Wasmtime implements this behavior.

NB: There is a documented [Linux bug][2] that pwrite does not respect
the offset and instead appends.  Wasmtime fixes this by tracking the
append flag status and hiding it from the OS.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/functions/write.html
[2]: https://linux.die.net/man/2/pwrite