Add support for test config overrides
It can be useful to override the config for certain tests. The main use
case for this is to override environment variables which control test
assertions (e.g. NO_DANGLING_FILESYSTEM, ERRNO_MODE_WINDOWS). When
running the tests in CI, it would be nicer to support this without
manually modifying the JSON test config files.