test runner: Report actual argv on errors
Instead of having the adapter files be top-level python scripts, now
they are modules with exported functions.  One of those functions
computes the argv to execute when running a test.

Also, clean up directories before and after each test, in addition to
before running the suite.  This makes one failure to clean up not cause
other cascading failures, and makes it more reliable to be able to run
test cases outside the test-runner harness.