conn-tests: Gracefully handle errors
This commit ensures that all possible errors, in the python driver
code and in the sockets program are handled gracefully, this allows
for a more robust testing.

Additionally this commit ensures that all tests are amended to take
into account the new structure of the `Output` class. The structure of
the `Output` class will most likely change as we move toward the `Op`
+ `Protocol` approach, therefore this commit doesn't introduce much
refactoring in that area.